requirements.txt
fastapi[all]
uvicorn
openai
aider-chat
elasticsearch
python-dotenv
pinecone-client